+++ Veriado.Infrastructure/FileSystem/IFilePathResolver.cs
@@ -6,6 +6,7 @@ public interface IFilePathResolver
 {
     string GetStorageRoot();
     string GetFullPath(string relativePath);
+    string GetFullPath(string relativePath, string? rootOverride);
     string GetFullPath(FileSystemEntity file);
     string GetRelativePath(string fullPath);
 }
